project_20_basic_langgraph_agent/requirements.txt
langchain
langchain-google-genai
python-dotenv
langgraph
uv